Split and move
