+++ jacodb-ets/src/main/kotlin/org/jacodb/ets/dsl/Node.kt
@@ -1,0 +1,44 @@
+/*
+ *  Copyright 2022 UnitTestBot contributors (utbot.org)
+ * <p>
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ * <p>
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.jacodb.ets.dsl
+
+sealed interface Node
+
+data object Nop : Node
+
+data class Assign(
+    val target: Local,
+    val expr: Expr,
+) : Node
+
+data class Return(
+    val expr: Expr,
+) : Node
+
+data class If(
+    val condition: Expr,
+    val thenBranch: List<Node>,
+    val elseBranch: List<Node>,
+) : Node
+
+data class Label(
+    val name: String,
+) : Node
+
+data class Goto(
+    val targetLabel: String,
+) : Node
